refactor: rename daily diary helpers
