+++ AGENTS.md
@@ -27,9 +27,11 @@ Travel Pet digital companion service that creates virtual pets for users and sen
 │   │   ├── generate-destination.prompt # Prompt for generating travel destinations
 │   │   └── generate-diary.prompt     # Prompt for generating diary entries
 │   ├── src/                  # Source code for Firebase Cloud Functions
-│   │   ├── createPetFlow.ts  # Pet creation logic (editable)
-│   │   ├── dailyDiaryFlow.ts # AI flows and helpers for diaries (editable)
-│   │   ├── diaryService.ts   # Destination & diary orchestration
+│   │   ├── createPetFlow.ts       # Pet creation logic (editable)
+│   │   ├── generateDestinationFlow.ts # Destination generation flow
+│   │   ├── generateDiaryFlow.ts  # Diary generation flow
+│   │   ├── diaryHelpers.ts     # Firestore and email helpers for diaries
+│   │   ├── diaryService.ts       # Destination & diary orchestration
 │   │   ├── genkit.config.ts  # Genkit plugin and flow configuration
 │   │   ├── config.ts         # Configuration settings (editable)
 │   │   ├── index.ts          # Main Cloud Functions entry point (editable)
