+++ functions/src/diaryHelpers.ts
@@ -2,55 +2,7 @@ import { sendEmail } from "./email";
 import { PetProfile, DiaryEntry } from "./types";
 import { db } from "./firebase";
 import { FieldValue } from "firebase-admin/firestore";
-import {
-  ai,
-  DestinationSchema,
-  GenerateDestinationInputSchema,
-  GenerateDiaryInputSchema,
-  DiarySchema,
-  Destination,
-  PetProfileData,
-} from "./genkit.config";
-
-const generateDestinationPrompt = ai.prompt<
-  typeof GenerateDestinationInputSchema,
-  typeof DestinationSchema
->("generate-destination");
-
-const generateDiaryPrompt = ai.prompt<
-  typeof GenerateDiaryInputSchema,
-  typeof DiarySchema
->("generate-diary");
-
-export const generateDestinationFlow = ai.defineFlow(
-  {
-    name: "generateDestinationFlow",
-    inputSchema: GenerateDestinationInputSchema,
-    outputSchema: DestinationSchema,
-  },
-  async (input) => {
-    const { output } = await generateDestinationPrompt(input);
-    if (!output) {
-      throw new Error("Failed to generate destination");
-    }
-    return output;
-  }
-);
-
-export const generateDiaryFlow = ai.defineFlow(
-  {
-    name: "generateDiaryFlow",
-    inputSchema: GenerateDiaryInputSchema,
-    outputSchema: DiarySchema,
-  },
-  async (input) => {
-    const { output } = await generateDiaryPrompt(input);
-    if (!output) {
-      throw new Error("Failed to generate diary");
-    }
-    return output;
-  }
-);
+import { Destination, PetProfileData } from "./genkit.config";
 
 
 // 分離されたFirestore読み取り関数
